Adding aggregating to mircrosoft
Microsoft constantly removes and adds the same job over and over again. Triggering the email notifications. To prevent this I added agregating similar to paypal. This takes in the new jobs and adds it to the list of all jobs. The previous way to do it was making all the current jobs as the list to be compared.